+++ tests/e2e/run_visible_browser_test.sh
@@ -31,9 +31,6 @@ if ! nc -z localhost 12000; then
   exit 1
 fi
 
-# Check if OpenHands is running
-echo "Connection to localhost port 12000 [tcp/entextxid] succeeded!"
-
 # Create test-results directory if it doesn't exist
 mkdir -p test-results
 
@@ -47,15 +44,23 @@ if [[ "$1" == *".py" ]]; then
   echo "Running: pytest -v --no-header --capture=no --no-headless --slow-mo=50 $1"
   poetry run pytest -v --no-header --capture=no --no-headless --slow-mo=50 "$1"
 elif [[ "$1" == "test_github_token_configuration" ]]; then
-  # Special case for the GitHub token configuration test
-  echo "Running: pytest -v --no-header --capture=no --no-headless --slow-mo=50 test_github_token_config.py"
-  poetry run pytest -v --no-header --capture=no --no-headless --slow-mo=50 test_github_token_config.py
+  # Run the GitHub token configuration test
+  echo "Running: pytest -v --no-header --capture=no --no-headless --slow-mo=50 test_e2e_workflow.py::test_github_token_configuration"
+  poetry run pytest -v --no-header --capture=no --no-headless --slow-mo=50 test_e2e_workflow.py::test_github_token_configuration
 elif [[ "$1" == "test_conversation_start" ]]; then
-  # Special case for the conversation start test
-  echo "Running: pytest -v --no-header --capture=no --no-headless --slow-mo=50 test_conversation_start.py"
-  poetry run pytest -v --no-header --capture=no --no-headless --slow-mo=50 test_conversation_start.py
+  # Run the conversation start test
+  echo "Running: pytest -v --no-header --capture=no --no-headless --slow-mo=50 test_e2e_workflow.py::test_conversation_start"
+  poetry run pytest -v --no-header --capture=no --no-headless --slow-mo=50 test_e2e_workflow.py::test_conversation_start
+elif [[ "$1" == "test_full_workflow" ]]; then
+  # Run the full workflow test
+  echo "Running: pytest -v --no-header --capture=no --no-headless --slow-mo=50 test_e2e_workflow.py::test_full_workflow"
+  poetry run pytest -v --no-header --capture=no --no-headless --slow-mo=50 test_e2e_workflow.py::test_full_workflow
+elif [[ "$1" == "test_simple_browser_navigation" ]]; then
+  # Run the simple browser navigation test
+  echo "Running: pytest -v --no-header --capture=no --no-headless --slow-mo=50 test_e2e_workflow.py::test_simple_browser_navigation"
+  poetry run pytest -v --no-header --capture=no --no-headless --slow-mo=50 test_e2e_workflow.py::test_simple_browser_navigation
 else
-  # It's a test name, run it from test_workflow.py
-  echo "Running: pytest -v --no-header --capture=no --no-headless --slow-mo=50 /Users/gneubig/work/OpenHands/tests/e2e/test_workflow.py::$1"
-  poetry run pytest -v --no-header --capture=no --no-headless --slow-mo=50 test_workflow.py::$1
+  # It's a test name, run it from test_e2e_workflow.py
+  echo "Running: pytest -v --no-header --capture=no --no-headless --slow-mo=50 test_e2e_workflow.py::$1"
+  poetry run pytest -v --no-header --capture=no --no-headless --slow-mo=50 test_e2e_workflow.py::$1
 fi
