--- tests/e2e/test_conversation_start.py
@@ -1,358 +1,0 @@
-"""
-Test for starting a conversation with the OpenHands agent.
-This test assumes the GitHub token is already configured.
-"""
-
-import os
-import time
-
-from playwright.sync_api import expect
-
-
-def get_readme_line_count():
-    """Get the line count of the main README.md file for verification."""
-    readme_path = os.path.join(
-        os.path.dirname(os.path.dirname(os.path.dirname(__file__))), 'README.md'
-    )
-    print(f'Looking for README.md at: {readme_path}')
-
-    try:
-        with open(readme_path, 'r', encoding='utf-8') as f:
-            lines = f.readlines()
-            return len(lines)
-    except Exception as e:
-        print(f'Error reading README.md: {e}')
-        return 0
-
-
-def test_conversation_start(page):
-    """
-    Test starting a conversation with the OpenHands agent:
-    1. Navigate to OpenHands (assumes GitHub token is already configured)
-    2. Select the OpenHands repository
-    3. Click Launch
-    4. Wait for the agent to initialize
-    5. Ask a question about the README.md file
-    6. Verify the agent responds correctly
-    """
-    # Create test-results directory if it doesn't exist
-    os.makedirs('test-results', exist_ok=True)
-
-    expected_line_count = get_readme_line_count()
-    print(f'Expected README.md line count: {expected_line_count}')
-
-    # Navigate to the OpenHands application
-    print('Step 1: Navigating to OpenHands application...')
-    page.goto('http://localhost:12000')
-    page.wait_for_load_state('networkidle', timeout=30000)
-
-    # Take initial screenshot
-    page.screenshot(path='test-results/conv_01_initial_load.png')
-    print('Screenshot saved: conv_01_initial_load.png')
-
-    # Handle any initial modals that might appear
-    try:
-        # Check for AI Provider Configuration modal
-        config_modal = page.locator('text=AI Provider Configuration')
-        if config_modal.is_visible(timeout=5000):
-            print('AI Provider Configuration modal detected')
-            save_button = page.locator('button:has-text("Save")')
-            if save_button.is_visible():
-                save_button.click()
-                page.wait_for_timeout(2000)
-
-        # Check for Privacy Preferences modal
-        privacy_modal = page.locator('text=Your Privacy Preferences')
-        if privacy_modal.is_visible(timeout=5000):
-            print('Privacy Preferences modal detected')
-            confirm_button = page.locator('button:has-text("Confirm Preferences")')
-            if confirm_button.is_visible():
-                confirm_button.click()
-                page.wait_for_timeout(2000)
-    except Exception as e:
-        print(f'Error handling initial modals: {e}')
-
-    # Step 2: Select the OpenHands repository
-    print('Step 2: Selecting openhands-agent/OpenHands repository...')
-
-    # Wait for the home screen to load
-    home_screen = page.locator('[data-testid="home-screen"]')
-    expect(home_screen).to_be_visible(timeout=15000)
-    print('Home screen is visible')
-
-    # Look for the repository dropdown/selector
-    repo_dropdown = page.locator('[data-testid="repo-dropdown"]')
-    expect(repo_dropdown).to_be_visible(timeout=5000)
-    print('Repository dropdown is visible')
-
-    # Click on the repository input to open dropdown
-    repo_dropdown.click()
-    page.wait_for_timeout(1000)
-
-    # Type the repository name
-    try:
-        # Try keyboard input for React Select component
-        page.keyboard.press('Control+a')  # Select all
-        page.keyboard.type('openhands-agent/OpenHands')
-        print('Used keyboard.type() for React Select component')
-    except Exception as e:
-        print(f'Keyboard input failed: {e}')
-
-    page.wait_for_timeout(2000)  # Wait for search results
-
-    # Try to find and click the repository option
-    option_selectors = [
-        '[data-testid="repo-dropdown"] [role="option"]:has-text("openhands-agent/OpenHands")',
-        '[data-testid="repo-dropdown"] [role="option"]:has-text("OpenHands")',
-        '[data-testid="repo-dropdown"] div[id*="option"]:has-text("openhands-agent/OpenHands")',
-        '[data-testid="repo-dropdown"] div[id*="option"]:has-text("OpenHands")',
-        '[role="option"]:has-text("openhands-agent/OpenHands")',
-        '[role="option"]:has-text("OpenHands")',
-        'div:has-text("openhands-agent/OpenHands"):not([id="aria-results"])',
-        'div:has-text("OpenHands"):not([id="aria-results"])',
-    ]
-
-    option_found = False
-    for selector in option_selectors:
-        try:
-            option = page.locator(selector).first
-            if option.is_visible(timeout=3000):
-                print(f'Found repository option with selector: {selector}')
-                try:
-                    option.click(force=True)
-                    print('Successfully clicked option with force=True')
-                    option_found = True
-                    page.wait_for_timeout(2000)
-                    break
-                except Exception:
-                    continue
-        except Exception:
-            continue
-
-    if not option_found:
-        print(
-            'Could not find repository option in dropdown, trying keyboard navigation'
-        )
-        page.keyboard.press('ArrowDown')
-        page.wait_for_timeout(500)
-        page.keyboard.press('Enter')
-        print('Used keyboard navigation to select option')
-
-    page.screenshot(path='test-results/conv_02_repo_selected.png')
-    print('Screenshot saved: conv_02_repo_selected.png')
-
-    # Step 3: Click Launch button
-    print('Step 3: Clicking Launch button...')
-
-    # Use the specific repository launch button
-    launch_button = page.locator('[data-testid="repo-launch-button"]')
-    expect(launch_button).to_be_visible(timeout=10000)
-
-    # Wait for the button to be enabled (not disabled)
-    max_wait_attempts = 30
-    button_enabled = False
-
-    for attempt in range(max_wait_attempts):
-        try:
-            is_disabled = launch_button.is_disabled()
-            if not is_disabled:
-                print(
-                    f'Repository Launch button is now enabled (attempt {attempt + 1})'
-                )
-                button_enabled = True
-                break
-            else:
-                print(
-                    f'Launch button still disabled, waiting... (attempt {attempt + 1}/{max_wait_attempts})'
-                )
-                page.wait_for_timeout(2000)
-        except Exception as e:
-            print(f'Error checking button state (attempt {attempt + 1}): {e}')
-            page.wait_for_timeout(2000)
-
-    # Try to click the button
-    try:
-        if button_enabled:
-            launch_button.click()
-            print('Launch button clicked normally')
-        else:
-            print('Launch button still disabled, trying JavaScript force click...')
-            # Use JavaScript to force click the button
-            result = page.evaluate("""() => {
-                const button = document.querySelector('[data-testid="repo-launch-button"]');
-                if (button) {
-                    console.log('Found button, removing disabled attribute');
-                    button.removeAttribute('disabled');
-                    console.log('Clicking button');
-                    button.click();
-                    return true;
-                }
-                return false;
-            }""")
-
-            if result:
-                print('Successfully force-clicked Launch button with JavaScript')
-            else:
-                print('JavaScript could not find the Launch button')
-    except Exception as e:
-        print(f'Error clicking Launch button: {e}')
-        page.screenshot(path='test-results/conv_03_launch_error.png')
-        print('Screenshot saved: conv_03_launch_error.png')
-        raise
-
-    # Step 4: Wait for conversation interface to load
-    print('Step 4: Waiting for conversation interface to load...')
-
-    # Wait for navigation to conversation page
-    navigation_timeout = 300000  # 5 minutes (300 seconds)
-    check_interval = 10000  # Check every 10 seconds
-
-    # Take a screenshot after clicking Launch
-    page.screenshot(path='test-results/conv_04_after_launch.png')
-    print('Screenshot saved: conv_04_after_launch.png')
-
-    # Wait for the conversation interface to appear
-    start_time = time.time()
-    conversation_loaded = False
-
-    while time.time() - start_time < navigation_timeout / 1000:
-        try:
-            # Check for conversation interface elements
-            conversation_interface = page.locator(
-                '[data-testid="conversation-interface"]'
-            )
-            if conversation_interface.is_visible(timeout=5000):
-                print('Conversation interface is visible')
-                conversation_loaded = True
-                break
-
-            # Alternative: Check for agent status indicators
-            agent_status = page.locator('[data-testid="agent-status"]')
-            if agent_status.is_visible(timeout=5000):
-                print('Agent status indicator is visible')
-                conversation_loaded = True
-                break
-
-            # Alternative: Check for message input
-            message_input = page.locator('[data-testid="message-input"]')
-            if message_input.is_visible(timeout=5000):
-                print('Message input is visible')
-                conversation_loaded = True
-                break
-
-            # Take periodic screenshots during the wait
-            if (time.time() - start_time) % (check_interval / 1000) < 1:
-                elapsed = int(time.time() - start_time)
-                page.screenshot(path=f'test-results/conv_05_waiting_{elapsed}s.png')
-                print(f'Screenshot saved: conv_05_waiting_{elapsed}s.png')
-
-            # Wait before checking again
-            page.wait_for_timeout(5000)
-
-        except Exception as e:
-            print(f'Error checking for conversation interface: {e}')
-            page.wait_for_timeout(5000)
-
-    if not conversation_loaded:
-        print('Timed out waiting for conversation interface to load')
-        page.screenshot(path='test-results/conv_06_timeout.png')
-        print('Screenshot saved: conv_06_timeout.png')
-        raise Exception('Timed out waiting for conversation interface to load')
-
-    # Step 5: Wait for agent to initialize
-    print('Step 5: Waiting for agent to initialize...')
-
-    # Wait for the agent to be ready for input
-    try:
-        # Look for "Agent is waiting for user input..." or similar message
-        ready_message = page.locator('text=Agent is waiting for user input')
-        expect(ready_message).to_be_visible(timeout=60000)  # Wait up to 1 minute
-        print('Agent is ready for input')
-    except Exception as e:
-        print(f'Could not confirm agent is ready: {e}')
-        # Continue anyway, as the message might be different
-
-    page.screenshot(path='test-results/conv_07_agent_ready.png')
-    print('Screenshot saved: conv_07_agent_ready.png')
-
-    # Step 6: Ask a question about the README.md file
-    print('Step 6: Asking question about README.md file...')
-
-    # Find the message input field
-    message_input = page.locator('[data-testid="message-input"], textarea')
-    expect(message_input).to_be_visible(timeout=10000)
-
-    # Type the question
-    message_input.fill('How many lines are there in the main README.md file?')
-    print('Entered question about README.md line count')
-
-    # Find and click the send button
-    send_button = page.locator('[data-testid="send-button"], button:has-text("Send")')
-    expect(send_button).to_be_visible(timeout=5000)
-    send_button.click()
-    print('Clicked send button')
-
-    page.screenshot(path='test-results/conv_08_question_sent.png')
-    print('Screenshot saved: conv_08_question_sent.png')
-
-    # Step 7: Wait for and verify the agent's response
-    print('Step 7: Waiting for agent response...')
-
-    # Wait for the agent to process the question
-    try:
-        # Look for "Agent is running task" or similar message
-        running_message = page.locator('text=Agent is running task')
-        expect(running_message).to_be_visible(timeout=30000)
-        print('Agent is processing the question')
-    except Exception as e:
-        print(f'Could not confirm agent is processing: {e}')
-        # Continue anyway, as the message might be different
-
-    # Wait for the agent to finish
-    try:
-        # Look for "Agent has finished the task" or similar message
-        finished_message = page.locator('text=Agent has finished')
-        expect(finished_message).to_be_visible(timeout=120000)  # Wait up to 2 minutes
-        print('Agent has finished processing')
-    except Exception as e:
-        print(f'Could not confirm agent has finished: {e}')
-        # Continue anyway, as the message might be different
-
-    page.screenshot(path='test-results/conv_09_agent_response.png')
-    print('Screenshot saved: conv_09_agent_response.png')
-
-    # Step 8: Verify the response contains the correct line count
-    print('Step 8: Verifying agent response...')
-
-    # Wait a bit more for the full response to be rendered
-    page.wait_for_timeout(5000)
-
-    # Get all message content
-    messages = page.locator('.message-content, [data-testid="message-content"]').all()
-
-    # Look for the line count in the last message
-    response_found = False
-    for message in reversed(messages):
-        try:
-            content = message.text_content()
-            print(f'Message content: {content}')
-
-            # Check if the message contains the line count
-            if str(expected_line_count) in content and 'README.md' in content:
-                print(
-                    f'✅ Agent correctly reported the README.md line count: {expected_line_count}'
-                )
-                response_found = True
-                break
-        except Exception as e:
-            print(f'Error checking message content: {e}')
-
-    if not response_found:
-        print('⚠️ Could not verify agent response contains correct line count')
-
-    # Final screenshot
-    page.screenshot(path='test-results/conv_10_test_complete.png')
-    print('Screenshot saved: conv_10_test_complete.png')
-
-    # Test passed if we got this far
-    print('Conversation test completed successfully')
