--- tests/e2e/test_workflow.py
@@ -1,1201 +1,0 @@
-import os
-
-import pytest
-from playwright.sync_api import Page, expect
-
-
-def get_readme_line_count():
-    """Get the line count of the README.md file."""
-    # Get the path to the repository root directory
-    current_dir = os.getcwd()
-    # If we're in the tests/e2e directory, go up two levels to the repo root
-    if current_dir.endswith('tests/e2e'):
-        repo_root = os.path.abspath(os.path.join(current_dir, '../..'))
-    else:
-        # If we're already at the repo root or somewhere else, try to find README.md
-        repo_root = current_dir
-
-    readme_path = os.path.join(repo_root, 'README.md')
-    print(f'Looking for README.md at: {readme_path}')
-    with open(readme_path, 'r', encoding='utf-8') as f:
-        lines = f.readlines()
-    return len(lines)
-
-
-@pytest.fixture(scope='module')
-def openhands_app():
-    """
-    Fixture that assumes OpenHands is already running on localhost.
-
-    This fixture checks if the OpenHands application is running on the expected port
-    and raises an exception if it's not available.
-    """
-    print('Checking if OpenHands is running...')
-
-    # Check if the application is running by trying to connect to the frontend port
-    max_attempts = 3
-    for attempt in range(1, max_attempts + 1):
-        try:
-            import socket
-            import time
-
-            s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            s.settimeout(2)
-            result = s.connect_ex(('localhost', 12000))
-            s.close()
-
-            if result == 0:
-                print(
-                    f'OpenHands is running on port 12000 (attempt {attempt}/{max_attempts})'
-                )
-                # Verify we can get HTML content
-                import urllib.request
-
-                try:
-                    with urllib.request.urlopen(
-                        'http://localhost:12000', timeout=5
-                    ) as response:
-                        html = response.read().decode('utf-8')
-                        if '<html' in html:
-                            print('Successfully received HTML content from OpenHands')
-                            yield None  # Success - yield the fixture value
-                            return
-                        else:
-                            print(
-                                f'WARNING: Port 12000 is open but not serving HTML content (attempt {attempt}/{max_attempts})'
-                            )
-                except Exception as e:
-                    print(
-                        f'WARNING: Port 12000 is open but could not fetch HTML: {e} (attempt {attempt}/{max_attempts})'
-                    )
-            else:
-                print(
-                    f'WARNING: OpenHands is not running on port 12000 (attempt {attempt}/{max_attempts})'
-                )
-
-            if attempt < max_attempts:
-                print('Waiting 5 seconds before retry...')
-                time.sleep(5)
-        except Exception as e:
-            print(f'ERROR checking OpenHands: {e} (attempt {attempt}/{max_attempts})')
-            if attempt < max_attempts:
-                print('Waiting 5 seconds before retry...')
-                time.sleep(5)
-
-    # If we get here, all attempts failed
-    raise Exception(
-        'OpenHands is not running on port 12000. Make sure to run "make run" before running the tests.'
-    )
-
-
-def test_readme_line_count():
-    """Test that we can count the lines in the README.md file."""
-    line_count = get_readme_line_count()
-    print(f'README.md has {line_count} lines')
-    assert line_count > 0, 'README.md should have at least one line'
-
-
-def test_simple_browser_navigation(page: Page):
-    """Test that we can navigate to a page using Playwright."""
-    # Navigate to the GitHub repository
-    page.goto('https://github.com/openhands-agent/OpenHands')
-
-    # Check that the page title contains "OpenHands"
-    expect(page).to_have_title(
-        'GitHub - openhands-agent/OpenHands: 🙌 OpenHands: Code Less, Make More'
-    )
-
-    # Check that the repository name is displayed
-    repo_header = page.locator('strong[itemprop="name"] a')
-    expect(repo_header).to_contain_text('OpenHands')
-
-    print('Successfully navigated to the OpenHands GitHub repository')
-
-
-def test_openhands_full_workflow(page, openhands_app):
-    """
-    Test the complete OpenHands workflow:
-    0. Assume environment variables are set (GITHUB_TOKEN, LLM_MODEL, LLM_API_KEY, LLM_BASE_URL)
-    1. Start OpenHands (already done by openhands_app fixture)
-    2. Use playwright to:
-       a. Handle AI Provider Configuration if needed
-       b. Click on the "openhands-agent/OpenHands" repo in the "Select a repo" dropdown
-       c. Click "Launch"
-       d. Check that the interface changes to the agent control interface
-       e. Check agent states: "Connecting", "Initializing Agent", "Agent is waiting for user input..."
-       f. Enter "How many lines are there in the main README.md file?" and submit
-       g. Check agent states: "Agent is running task" and "Agent has finished the task."
-       h. Check that the final agent message contains the correct line count
-    """
-    # Create test-results directory if it doesn't exist
-    os.makedirs('test-results', exist_ok=True)
-
-    expected_line_count = get_readme_line_count()
-    print(f'Expected README.md line count: {expected_line_count}')
-
-    # Navigate to the OpenHands application
-    print('Step 1: Navigating to OpenHands application...')
-    page.goto('http://localhost:12000')
-    page.wait_for_load_state('networkidle', timeout=30000)
-
-    # Take initial screenshot
-    page.screenshot(path='test-results/01_initial_load.png')
-    print('Screenshot saved: 01_initial_load.png')
-
-    # Step 2a: Handle AI Provider Configuration if it appears
-    print('Step 2a: Checking for AI Provider Configuration modal...')
-    try:
-        # Check if the AI Provider Configuration modal is present
-        config_modal = page.locator('text=AI Provider Configuration')
-        if config_modal.is_visible(timeout=5000):
-            print('AI Provider Configuration modal detected')
-
-            # Check if API key field is empty and fill it if needed
-            api_key_input = page.locator(
-                'input[placeholder*="API"], input[name*="api"], input[type="password"]'
-            ).first
-            if api_key_input.is_visible():
-                current_value = api_key_input.input_value()
-                if not current_value.strip():
-                    print('API key field is empty, filling with environment variable')
-                    # Use a placeholder API key for testing
-                    api_key_input.fill('test-api-key-from-env')
-                else:
-                    print('API key field already has a value')
-
-            # Click Save button
-            save_button = page.locator('button:has-text("Save")')
-            if save_button.is_visible():
-                print('Clicking Save button in AI Provider Configuration')
-                save_button.click()
-                page.wait_for_timeout(2000)  # Wait for modal to close
-
-        page.screenshot(path='test-results/02_after_config.png')
-        print('Screenshot saved: 02_after_config.png')
-
-    except Exception as e:
-        print(f'No AI Provider Configuration modal found or error handling it: {e}')
-
-    # Step 2b: Handle Privacy Preferences modal if it appears
-    print('Step 2b: Checking for Privacy Preferences modal...')
-    try:
-        # Check if the Privacy Preferences modal is present
-        privacy_modal = page.locator('text=Your Privacy Preferences')
-        if privacy_modal.is_visible(timeout=5000):
-            print('Privacy Preferences modal detected')
-
-            # Click "Confirm Preferences" button
-            confirm_button = page.locator('button:has-text("Confirm Preferences")')
-            if confirm_button.is_visible():
-                print('Clicking Confirm Preferences button')
-                confirm_button.click()
-                page.wait_for_timeout(2000)  # Wait for modal to close
-            else:
-                print(
-                    'Confirm Preferences button not found, looking for alternatives...'
-                )
-                # Try other possible button texts
-                alt_buttons = [
-                    'button:has-text("Confirm")',
-                    'button:has-text("Accept")',
-                    'button:has-text("OK")',
-                    'button[type="submit"]',
-                ]
-                for button_selector in alt_buttons:
-                    try:
-                        button = page.locator(button_selector)
-                        if button.is_visible(timeout=1000):
-                            print(f'Found alternative button: {button_selector}')
-                            button.click()
-                            page.wait_for_timeout(2000)
-                            break
-                    except Exception:
-                        continue
-
-        page.screenshot(path='test-results/03_after_privacy.png')
-        print('Screenshot saved: 03_after_privacy.png')
-
-    except Exception as e:
-        print(f'No Privacy Preferences modal found or error handling it: {e}')
-
-    # Step 2c: Check if GitHub token is already configured or needs to be set
-    print('Step 2c: Checking if GitHub token is configured...')
-    page.screenshot(path='test-results/04_before_token_check.png')
-    print('Screenshot saved: 04_before_token_check.png')
-
-    try:
-        # First, check if we're already on the home screen with repository selection
-        # This means the GitHub token is already configured in ~/.openhands/settings.json
-        connect_to_provider = page.locator('text=Connect to a Repository')
-
-        if connect_to_provider.is_visible(timeout=3000):
-            print('Found "Connect to a Repository" section')
-
-            # Check if we need to configure a provider (GitHub token)
-            navigate_to_settings_button = page.locator(
-                '[data-testid="navigate-to-settings-button"]'
-            )
-
-            if navigate_to_settings_button.is_visible(timeout=3000):
-                print('GitHub token not configured. Need to navigate to settings...')
-
-                # Click the Settings button to navigate to the settings page
-                navigate_to_settings_button.click()
-                page.wait_for_load_state('networkidle', timeout=10000)
-                page.wait_for_timeout(3000)  # Wait for navigation to complete
-
-                # We should now be on the /settings/integrations page
-                print('Navigated to settings page, looking for GitHub token input...')
-
-                # Check if we're on the settings page with the integrations tab
-                settings_screen = page.locator('[data-testid="settings-screen"]')
-                if settings_screen.is_visible(timeout=5000):
-                    print('Settings screen is visible')
-
-                    # Make sure we're on the Integrations tab
-                    integrations_tab = page.locator('text=Integrations')
-                    if integrations_tab.is_visible(timeout=3000):
-                        # Check if we need to click the tab
-                        if not page.url.endswith('/settings/integrations'):
-                            print('Clicking Integrations tab...')
-                            integrations_tab.click()
-                            page.wait_for_load_state('networkidle')
-                            page.wait_for_timeout(2000)
-
-                    # Now look for the GitHub token input
-                    github_token_input = page.locator(
-                        '[data-testid="github-token-input"]'
-                    )
-                    if github_token_input.is_visible(timeout=5000):
-                        print('Found GitHub token input field')
-
-                        # Fill in the GitHub token from environment variable
-                        github_token = os.getenv('GITHUB_TOKEN', '')
-                        if github_token:
-                            # Clear the field first, then fill it
-                            github_token_input.clear()
-                            github_token_input.fill(github_token)
-                            print(
-                                f'Filled GitHub token from environment variable (length: {len(github_token)})'
-                            )
-
-                            # Verify the token was filled
-                            filled_value = github_token_input.input_value()
-                            if filled_value:
-                                print(
-                                    f'Token field now contains value of length: {len(filled_value)}'
-                                )
-                            else:
-                                print(
-                                    'WARNING: Token field appears to be empty after filling'
-                                )
-
-                            # Look for the Save Changes button and ensure it's enabled
-                            save_button = page.locator('[data-testid="submit-button"]')
-                            if save_button.is_visible(timeout=3000):
-                                # Check if button is enabled
-                                is_disabled = save_button.is_disabled()
-                                print(
-                                    f'Save Changes button found, disabled: {is_disabled}'
-                                )
-
-                                if not is_disabled:
-                                    print('Clicking Save Changes button...')
-                                    save_button.click()
-
-                                    # Wait for the save operation to complete
-                                    try:
-                                        # Wait for the button to show "Saving..." (if it does)
-                                        page.wait_for_timeout(1000)
-
-                                        # Wait for the save to complete - button should be disabled again
-                                        page.wait_for_function(
-                                            'document.querySelector(\'[data-testid="submit-button"]\').disabled === true',
-                                            timeout=10000,
-                                        )
-                                        print(
-                                            'Save operation completed - form is now clean'
-                                        )
-                                    except Exception:
-                                        print(
-                                            'Save operation completed (timeout waiting for form clean state)'
-                                        )
-
-                                    # Navigate back to home page after successful save
-                                    print('Navigating back to home page...')
-                                    page.goto('http://localhost:12000')
-                                    page.wait_for_load_state('networkidle')
-                                    page.wait_for_timeout(
-                                        5000
-                                    )  # Wait longer for providers to be updated
-                                else:
-                                    print(
-                                        'Save Changes button is disabled - form may be invalid'
-                                    )
-                            else:
-                                print('Save Changes button not found')
-                        else:
-                            print('No GitHub token found in environment variables')
-                    else:
-                        print('GitHub token input field not found on settings page')
-                        # Take a screenshot to see what's on the page
-                        page.screenshot(path='test-results/04b_settings_debug.png')
-                        print('Debug screenshot saved: 04b_settings_debug.png')
-                else:
-                    print('Settings screen not found')
-            else:
-                # Branch 2: GitHub token is already configured, repository selection is available
-                print(
-                    'GitHub token is already configured, repository selection is available'
-                )
-
-                # Check if we need to update the token by going to settings manually
-                settings_button = page.locator('button:has-text("Settings")')
-                if settings_button.is_visible(timeout=3000):
-                    print(
-                        'Settings button found, clicking to navigate to settings page...'
-                    )
-                    settings_button.click()
-                    page.wait_for_load_state('networkidle', timeout=10000)
-                    page.wait_for_timeout(3000)  # Wait for navigation to complete
-
-                    # Navigate to the Integrations tab
-                    integrations_tab = page.locator('text=Integrations')
-                    if integrations_tab.is_visible(timeout=3000):
-                        print('Clicking Integrations tab...')
-                        integrations_tab.click()
-                        page.wait_for_load_state('networkidle')
-                        page.wait_for_timeout(2000)
-
-                        # Now look for the GitHub token input
-                        github_token_input = page.locator(
-                            '[data-testid="github-token-input"]'
-                        )
-                        if github_token_input.is_visible(timeout=5000):
-                            print('Found GitHub token input field')
-
-                            # Fill in the GitHub token from environment variable
-                            github_token = os.getenv('GITHUB_TOKEN', '')
-                            if github_token:
-                                # Clear the field first, then fill it
-                                github_token_input.clear()
-                                github_token_input.fill(github_token)
-                                print(
-                                    f'Filled GitHub token from environment variable (length: {len(github_token)})'
-                                )
-
-                                # Look for the Save Changes button and ensure it's enabled
-                                save_button = page.locator(
-                                    '[data-testid="submit-button"]'
-                                )
-                                if (
-                                    save_button.is_visible(timeout=3000)
-                                    and not save_button.is_disabled()
-                                ):
-                                    print('Clicking Save Changes button...')
-                                    save_button.click()
-                                    page.wait_for_timeout(3000)
-
-                                # Navigate back to home page
-                                print('Navigating back to home page...')
-                                page.goto('http://localhost:12000')
-                                page.wait_for_load_state('networkidle')
-                                page.wait_for_timeout(3000)
-                        else:
-                            print(
-                                'GitHub token input field not found, going back to home page'
-                            )
-                            page.goto('http://localhost:12000')
-                            page.wait_for_load_state('networkidle')
-                    else:
-                        print('Integrations tab not found, going back to home page')
-                        page.goto('http://localhost:12000')
-                        page.wait_for_load_state('networkidle')
-                else:
-                    print('Settings button not found, continuing with existing token')
-        else:
-            print('Could not find "Connect to a Repository" section')
-
-        page.screenshot(path='test-results/04_after_settings.png')
-        print('Screenshot saved: 04_after_settings.png')
-
-    except Exception as e:
-        print(f'Error checking GitHub token configuration: {e}')
-        page.screenshot(path='test-results/04_settings_error.png')
-        print('Screenshot saved: 04_settings_error.png')
-
-    # Step 2d: Wait for home screen and find the repository selector
-    print('Step 2d: Looking for repository selector...')
-
-    # Wait for the home screen to load
-    home_screen = page.locator('[data-testid="home-screen"]')
-    expect(home_screen).to_be_visible(timeout=15000)
-    print('Home screen is visible')
-
-    # Look for the repository dropdown/selector
-    # Try multiple possible selectors for the repository dropdown
-    repo_selectors = [
-        '[data-testid="repo-dropdown"]',  # React Select async dropdown
-        'input[placeholder*="Search repositories"]',
-        'input[placeholder*="repository"]',
-        '[data-testid*="repo"]',
-        'button:has-text("Select")',
-        '.repository-selector',
-        '[role="combobox"]',
-    ]
-
-    repo_input = None
-    for selector in repo_selectors:
-        try:
-            element = page.locator(selector).first
-            if element.is_visible(timeout=2000):
-                repo_input = element
-                print(f'Found repository selector with: {selector}')
-                break
-        except Exception:
-            continue
-
-    if not repo_input:
-        print('Repository selector not found, taking screenshot for debugging')
-        page.screenshot(path='test-results/05_repo_selector_not_found.png')
-        # Try to find any input or dropdown elements
-        all_inputs = page.locator('input, select, [role="combobox"], [role="textbox"]')
-        count = all_inputs.count()
-        print(f'Found {count} input/dropdown elements on the page')
-        for i in range(min(count, 5)):  # Check first 5 elements
-            element = all_inputs.nth(i)
-            placeholder = element.get_attribute('placeholder') or ''
-            role = element.get_attribute('role') or ''
-            print(f'  Element {i}: placeholder="{placeholder}", role="{role}"')
-
-        raise Exception('Could not find repository selector')
-
-    # Step 2e: Select the OpenHands repository
-    print('Step 2e: Selecting openhands-agent/OpenHands repository...')
-
-    # Click on the repository input to open dropdown
-    repo_input.click()
-    page.wait_for_timeout(1000)
-
-    # For React Select components, we need to type directly after clicking
-    # The input field appears after clicking the dropdown
-    try:
-        # Try to fill first (for regular inputs)
-        repo_input.fill('openhands-agent/OpenHands')
-        print('Used fill() method for repository input')
-    except Exception as e:
-        print(f'Fill failed ({e}), trying keyboard input for React Select...')
-        # If fill fails, this is likely a React Select component
-        # Clear any existing text and type
-        page.keyboard.press('Control+a')  # Select all
-        page.keyboard.type('openhands-agent/OpenHands')
-        print('Used keyboard.type() for React Select component')
-
-    page.wait_for_timeout(2000)  # Wait for search results
-
-    # Check if the repository is already selected in the dropdown
-    # The React Select component might auto-select the first matching result
-    try:
-        # Look for the selected value in the React Select component
-        selected_repo = page.locator('[data-testid="repo-dropdown"] input')
-        if selected_repo.is_visible(timeout=2000):
-            selected_value = selected_repo.input_value()
-            print(f'Repository dropdown input value: "{selected_value}"')
-            if (
-                'openhands-agent/OpenHands' in selected_value
-                or 'OpenHands' in selected_value
-            ):
-                print('openhands-agent/OpenHands repository is already selected')
-            else:
-                print(f'Different repository in input: {selected_value}')
-        else:
-            pass
-    except Exception:
-        pass
-
-    # Even if the text is in the input, we need to click the dropdown option to complete selection
-    # Look for the OpenHands repository in the dropdown options
-    print(
-        'Looking for OpenHands repository in dropdown options to complete selection...'
-    )
-
-    # Wait a bit more for the dropdown to populate
-    page.wait_for_timeout(1000)
-
-    # Try to find and click the repository option
-    option_found = False
-
-    # React Select creates options with specific structure - target the actual clickable option
-    option_selectors = [
-        # React Select option selectors (most specific first)
-        '[data-testid="repo-dropdown"] [role="option"]:has-text("openhands-agent/OpenHands")',
-        '[data-testid="repo-dropdown"] [role="option"]:has-text("OpenHands")',
-        '[data-testid="repo-dropdown"] div[id*="option"]:has-text("openhands-agent/OpenHands")',
-        '[data-testid="repo-dropdown"] div[id*="option"]:has-text("OpenHands")',
-        # Generic option selectors
-        '[role="option"]:has-text("openhands-agent/OpenHands")',
-        '[role="option"]:has-text("OpenHands")',
-        # Fallback selectors (but avoid aria-results span)
-        'div:has-text("openhands-agent/OpenHands"):not([id="aria-results"])',
-        'div:has-text("OpenHands"):not([id="aria-results"])',
-    ]
-
-    for selector in option_selectors:
-        try:
-            option = page.locator(selector).first
-            if option.is_visible(timeout=3000):
-                print(f'Found repository option with selector: {selector}')
-                # Try force click first to bypass element interception
-                try:
-                    option.click(force=True)
-                    print('Successfully clicked option with force=True')
-                    option_found = True
-                    page.wait_for_timeout(
-                        2000
-                    )  # Wait longer for React Select to update
-                    break
-                except Exception as force_error:
-                    print(f'Force click failed: {force_error}, trying regular click...')
-                    option.click()
-                    option_found = True
-                    page.wait_for_timeout(
-                        2000
-                    )  # Wait longer for React Select to update
-                    break
-        except Exception as e:
-            print(f'Selector {selector} failed: {e}')
-            continue
-
-    if not option_found:
-        print('Could not find repository option in dropdown')
-        # Try keyboard navigation as fallback
-        print('Trying keyboard navigation: Arrow Down + Enter')
-        page.keyboard.press('ArrowDown')  # Navigate to first option
-        page.wait_for_timeout(500)
-        page.keyboard.press('Enter')  # Select the option
-        print('Used keyboard navigation to select option')
-        page.wait_for_timeout(1000)
-
-    # Verify that the repository selection actually worked
-    page.wait_for_timeout(1000)
-    print('Verifying repository selection...')
-    try:
-        # Check if the React Select shows the selected repository
-        selected_value_element = page.locator(
-            '[data-testid="repo-dropdown"] .css-1jcgswf'
-        )
-        if selected_value_element.is_visible(timeout=2000):
-            selected_text = selected_value_element.text_content()
-            print(f'React Select selected value: "{selected_text}"')
-            if 'OpenHands' in selected_text:
-                print('✅ Repository selection verified - React Select shows OpenHands')
-            else:
-                print(
-                    f'⚠️ Repository selection may have failed - shows: {selected_text}'
-                )
-        else:
-            print('⚠️ Could not find React Select selected value element')
-    except Exception as e:
-        print(f'Error verifying repository selection: {e}')
-
-    page.screenshot(path='test-results/07_repo_selected.png')
-    print('Screenshot saved: 07_repo_selected.png')
-
-    # Step 2f: Click Launch button (repository-specific)
-    print('Step 2f: Looking for repository Launch button...')
-
-    # Use the specific repository launch button, not the "Launch from Scratch" button
-    launch_button = page.locator('[data-testid="repo-launch-button"]')
-
-    # Wait for the button to be visible and enabled
-    print('Waiting for repository Launch button to be enabled...')
-    expect(launch_button).to_be_visible(timeout=10000)
-
-    # Wait for the button to be enabled (not disabled)
-    # The button should become enabled once repository selection is complete
-    launch_button.wait_for(state='attached', timeout=5000)
-
-    # Check if button is enabled by waiting for it to not have disabled attribute
-    # Increase wait time significantly since you mentioned it takes a long time
-    max_wait_attempts = 30  # Increased from 10 to 30 (30 seconds total)
-    button_enabled = False
-
-    for attempt in range(max_wait_attempts):
-        try:
-            # Check if button is enabled
-            is_disabled = launch_button.is_disabled()
-            if not is_disabled:
-                print(
-                    f'Repository Launch button is now enabled (attempt {attempt + 1})'
-                )
-                button_enabled = True
-                break
-            else:
-                print(
-                    f'Launch button still disabled, waiting... (attempt {attempt + 1}/{max_wait_attempts})'
-                )
-                page.wait_for_timeout(2000)  # Increased from 1000ms to 2000ms
-        except Exception as e:
-            print(f'Error checking button state (attempt {attempt + 1}): {e}')
-            page.wait_for_timeout(2000)  # Increased from 1000ms to 2000ms
-
-    if not button_enabled:
-        print(
-            'Launch button is still disabled after waiting, taking debug screenshot...'
-        )
-        page.screenshot(path='test-results/07b_launch_button_debug.png')
-        print('Debug screenshot saved: 07b_launch_button_debug.png')
-
-        # Try to proceed anyway - maybe the button will work
-        print('Attempting to click Launch button despite disabled state...')
-    else:
-        print('Launch button is enabled and ready to click')
-
-    # Verify the button is enabled before clicking (but don't fail if it's not)
-    try:
-        expect(launch_button).to_be_enabled()
-        print('Launch button verification passed')
-        launch_button.click()
-        print('Launch button clicked')
-    except Exception as e:
-        print(f'Launch button verification failed: {e}')
-        print('Proceeding with JavaScript force click...')
-
-        # Use JavaScript to force click the button
-        try:
-            result = page.evaluate("""() => {
-                const button = document.querySelector('[data-testid="repo-launch-button"]');
-                if (button) {
-                    console.log('Found button, removing disabled attribute');
-                    button.removeAttribute('disabled');
-                    console.log('Clicking button');
-                    button.click();
-                    return true;
-                }
-                return false;
-            }""")
-
-            if result:
-                print('Successfully force-clicked Launch button with JavaScript')
-            else:
-                print('JavaScript could not find the Launch button')
-
-            # Wait for any navigation or state change
-            page.wait_for_timeout(5000)
-        except Exception as js_error:
-            print(f'JavaScript force click failed: {js_error}')
-            # Last resort: try clicking with force=True option
-            try:
-                launch_button.click(force=True, timeout=5000)
-                print('Used force=True to click the Launch button')
-            except Exception as force_error:
-                print(f'Force click also failed: {force_error}')
-                print('Test may fail due to inability to click Launch button')
-
-    # Step 2g: Wait for conversation interface to load
-    print('Step 2g: Waiting for conversation interface to load...')
-
-    # After clicking Launch, the button shows "Loading..." and creates a conversation
-    # Wait for the Loading state to appear first
-    print('Waiting for Loading state to appear...')
-    loading_button = page.locator(
-        '[data-testid="repo-launch-button"]:has-text("Loading")'
-    )
-    try:
-        expect(loading_button).to_be_visible(timeout=10000)
-        print('Loading state detected on Launch button')
-    except Exception:
-        print('Loading state not detected, continuing...')
-
-    # Wait for navigation to conversation page
-    print('Waiting for navigation to conversation page...')
-
-    # Increase timeout to 5 minutes to account for Docker image building
-    navigation_timeout = 300000  # 5 minutes (300 seconds)
-    check_interval = 10000  # Check every 10 seconds
-
-    start_time = page.evaluate('Date.now()')
-    navigation_successful = False
-
-    while True:
-        try:
-            current_time = page.evaluate('Date.now()')
-            elapsed = current_time - start_time
-        except Exception as e:
-            # If execution context is destroyed, it likely means navigation is happening
-            if (
-                'Execution context was destroyed' in str(e)
-                or 'navigation' in str(e).lower()
-            ):
-                print(
-                    '🚀 Navigation detected (execution context destroyed) - checking URL...'
-                )
-                # Wait a moment for navigation to complete
-                page.wait_for_timeout(2000)
-                current_url = page.url
-                if '/conversations/' in current_url:
-                    print(
-                        f'✅ Successfully navigated to conversation page: {current_url}'
-                    )
-                    navigation_successful = True
-                    break
-                else:
-                    print(
-                        f'⚠️ Navigation detected but not to conversation page: {current_url}'
-                    )
-                    # Continue monitoring
-                    elapsed = 0  # Reset timer since we detected navigation activity
-            else:
-                print(f'⚠️ Unexpected error during navigation monitoring: {e}')
-                elapsed = 0  # Reset timer and continue
-
-        # Check if we've navigated to a conversation page
-        current_url = page.url
-        if '/conversations/' in current_url:
-            print(f'✅ Successfully navigated to conversation page: {current_url}')
-            navigation_successful = True
-            break
-
-        # Check if we've exceeded the timeout
-        if elapsed > navigation_timeout:
-            print(f'❌ Navigation timeout after {elapsed / 1000:.1f} seconds')
-            break
-
-        # Print periodic status updates
-        if elapsed % check_interval < 1000:  # Print roughly every 10 seconds
-            print(
-                f'⏳ Still waiting for navigation... ({elapsed / 1000:.1f}s elapsed, current URL: {current_url})'
-            )
-
-            # Check if Launch button still shows Loading
-            try:
-                loading_button = page.locator(
-                    '[data-testid="repo-launch-button"]:has-text("Loading")'
-                )
-                if loading_button.is_visible(timeout=1000):
-                    print(
-                        '   🔄 Launch button still shows "Loading..." (likely Docker build in progress)'
-                    )
-                else:
-                    launch_button = page.locator('[data-testid="repo-launch-button"]')
-                    if launch_button.is_visible(timeout=1000):
-                        button_text = launch_button.text_content()
-                        print(f'   📝 Launch button text: "{button_text}"')
-            except Exception:
-                pass
-
-            # Provide helpful context about what might be happening
-            if elapsed > 60000:  # After 1 minute
-                print(
-                    '   💡 Extended wait time is normal - Docker runtime image may be building'
-                )
-            if elapsed > 120000:  # After 2 minutes
-                print(
-                    '   🐳 Docker build can take 3-5 minutes on first run or with new dependencies'
-                )
-
-        # Wait a bit before next check
-        page.wait_for_timeout(1000)
-
-    if not navigation_successful:
-        current_url = page.url
-        print(f'Current URL after timeout: {current_url}')
-
-        # Take a screenshot to debug
-        page.screenshot(path='test-results/08_navigation_timeout.png')
-        print('Screenshot saved: 08_navigation_timeout.png')
-
-        # Check for error messages
-        error_selectors = [
-            'text=Error',
-            'text=Failed',
-            'text=Something went wrong',
-            '[role="alert"]',
-            '.error',
-            '.alert',
-            '[data-testid*="error"]',
-        ]
-
-        for selector in error_selectors:
-            try:
-                error_element = page.locator(selector).first
-                if error_element.is_visible(timeout=2000):
-                    error_text = error_element.text_content()
-                    print(f'Found error message: {error_text}')
-                    break
-            except Exception:
-                continue
-
-        # If navigation failed, we can't continue with the test
-        raise Exception(
-            f'Failed to navigate to conversation page after {navigation_timeout / 1000} seconds. Current URL: {current_url}'
-        )
-
-    # Wait for the conversation interface to fully load
-    print('Waiting for conversation interface to load...')
-    conversation_interface = page.locator('[data-testid="app-route"]')
-    expect(conversation_interface).to_be_visible(timeout=15000)
-    print('Conversation interface loaded successfully')
-
-    # Take a screenshot of the conversation interface
-    page.screenshot(path='test-results/08_conversation_loaded.png')
-    print('Screenshot saved: 08_conversation_loaded.png')
-
-    # Step 2h: Check agent initialization states
-    print('Step 2h: Monitoring agent states during initialization...')
-
-    # Wait for agent to initialize - look for various states
-    # The agent goes through: LOADING -> INIT -> AWAITING_USER_INPUT
-
-    # Wait for agent to be ready for user input
-    print('Waiting for agent to be ready for user input...')
-
-    # Look for the chat input field to be enabled (indicates agent is ready)
-    chat_input_selectors = [
-        'textarea[placeholder*="message"]',
-        'textarea[placeholder*="Message"]',
-        '[data-testid="message-input"]',
-        'textarea:not([disabled])',
-    ]
-
-    input_found = False
-    for selector in chat_input_selectors:
-        try:
-            input_element = page.locator(selector).first
-            if input_element.is_visible(timeout=30000):
-                print(f'Found chat input element: {selector}')
-                input_found = True
-                break
-        except Exception:
-            continue
-
-    if not input_found:
-        print('Chat input not found, taking screenshot for debugging...')
-        page.screenshot(path='test-results/09_no_chat_input.png')
-        print('Screenshot saved: 09_no_chat_input.png')
-        raise Exception('Chat input field not found - agent may not be ready')
-
-    # Take a screenshot showing the ready state
-    page.screenshot(path='test-results/09_agent_ready.png')
-    print('Screenshot saved: 09_agent_ready.png')
-
-    # Step 2i: Wait for the agent to be ready
-    print('Waiting for agent to be ready (this may take 1-2 minutes)...')
-
-    # Wait for the status to change from "Connecting..." to "Agent is awaiting user input"
-    # The status is displayed in the status bar with a specific CSS class
-    status_selector = (
-        'div.flex.items-center.bg-base-secondary span.text-sm.text-stone-400'
-    )
-
-    # Wait for up to 3 minutes (180000ms) for the agent to be ready
-    max_wait_time = 180000  # 3 minutes in milliseconds
-    status_element = page.locator(status_selector)
-
-    # First, make sure the status element is visible
-    expect(status_element).to_be_visible(timeout=30000)
-
-    # Then wait for the status to change to "Agent is awaiting user input"
-    print(
-        'Status element found, waiting for status to change to "Agent is awaiting user input"...'
-    )
-
-    # Define a function to check if the agent is ready
-    def check_agent_ready():
-        status_text = status_element.text_content()
-        print(f'Current status: {status_text}')
-        return 'awaiting user input' in status_text.lower() if status_text else False
-
-    # Poll until the agent is ready or timeout is reached
-    start_time = page.evaluate('() => Date.now()')
-    while True:
-        if check_agent_ready():
-            print('Agent is ready!')
-            break
-
-        # Check if we've exceeded the maximum wait time
-        current_time = page.evaluate('() => Date.now()')
-        elapsed_time = current_time - start_time
-        if elapsed_time > max_wait_time:
-            # Take a screenshot for debugging
-            page.screenshot(path='test-results/agent_not_ready_timeout.png')
-            raise Exception(
-                f'Timed out waiting for agent to be ready after {max_wait_time / 1000} seconds'
-            )
-
-        # Wait a bit before checking again
-        page.wait_for_timeout(2000)  # Wait 2 seconds between checks
-
-    # Take a screenshot showing the ready state
-    page.screenshot(path='test-results/10_agent_ready.png')
-    print('Screenshot saved: 10_agent_ready.png')
-
-    # Now proceed with entering the question
-    print('Step 2i: Entering question about README.md line count...')
-
-    question = 'How many lines are there in the main README.md file?'
-
-    # Find the chat input using multiple possible selectors
-    # The chat input might be different depending on the UI state
-    chat_input_selectors = [
-        'textarea[placeholder*="message"]',
-        'textarea[placeholder*="Message"]',
-        '[data-testid="message-input"]',
-        'textarea:not([disabled])',
-        'div[contenteditable="true"]',
-        '.chat-input textarea',
-        'form textarea',
-        'textarea',
-    ]
-
-    # Try each selector until we find a visible input
-    chat_input = None
-    for selector in chat_input_selectors:
-        try:
-            input_element = page.locator(selector).first
-            if input_element.is_visible(timeout=5000):
-                print(f'Found chat input element with selector: {selector}')
-                chat_input = input_element
-                break
-        except Exception as e:
-            print(f'Selector {selector} not found: {e}')
-            continue
-
-    if not chat_input:
-        # Take a screenshot for debugging
-        page.screenshot(path='test-results/chat_input_not_found.png')
-        print('Screenshot saved: chat_input_not_found.png')
-
-        # Try to find any visible elements to help debug
-        page.screenshot(path='test-results/debug_visible_elements.png')
-        print('Debug screenshot saved: debug_visible_elements.png')
-
-        # Check if there are any form elements
-        form_elements = page.locator('form').all()
-        print(f'Found {len(form_elements)} form elements')
-
-        # Check if there are any textarea elements
-        textarea_elements = page.locator('textarea').all()
-        print(f'Found {len(textarea_elements)} textarea elements')
-
-        raise Exception('Chat input field not found after trying multiple selectors')
-
-    print(f'Entering question: {question}')
-    chat_input.fill(question)
-
-    # Submit the message (look for submit button or press Enter)
-    submit_selectors = [
-        'button[type="submit"]',
-        'button:has-text("Submit")',
-        'button:has-text("Send")',
-        '[data-testid="submit-button"]',
-    ]
-
-    submitted = False
-    for selector in submit_selectors:
-        try:
-            submit_button = page.locator(selector).first
-            if submit_button.is_visible(timeout=5000):
-                print(f'Found submit button: {selector}')
-                submit_button.click()
-                submitted = True
-                break
-        except Exception:
-            continue
-
-    if not submitted:
-        # Try pressing Enter as fallback
-        print('Submit button not found, trying Enter key...')
-        chat_input.press('Enter')
-        submitted = True
-
-    print('Question submitted successfully')
-
-    # Take a screenshot after submitting
-    page.screenshot(path='test-results/10_question_submitted.png')
-    print('Screenshot saved: 10_question_submitted.png')
-
-    # Step 2j: Monitor agent states during task execution
-    print('Step 2j: Monitoring agent states during task execution...')
-
-    # Wait for agent to start processing (look for running state or typing indicator)
-    running_indicators = [
-        'text=running',
-        'text=Agent is running',
-        '[data-testid="typing-indicator"]',
-        '.animate-bounce',  # Typing indicator animation
-    ]
-
-    print('Waiting for agent to start processing...')
-    processing_started = False
-    for selector in running_indicators:
-        try:
-            element = page.locator(selector).first
-            if element.is_visible(timeout=15000):
-                print(f'Agent started processing - found: {selector}')
-                processing_started = True
-                break
-        except Exception:
-            continue
-
-    if not processing_started:
-        print('Agent processing indicators not found, continuing anyway...')
-
-    # Wait for agent to finish the task
-    print('Waiting for agent to complete the task...')
-
-    # Look for completion indicators or final response
-
-    # Wait for a reasonable amount of time for the agent to complete the task
-    # README.md line counting should be quick
-    page.wait_for_timeout(60000)  # Wait up to 60 seconds
-
-    # Take a screenshot of the final state
-    page.screenshot(path='test-results/11_task_completed.png')
-    print('Screenshot saved: 11_task_completed.png')
-
-    # Step 2k: Verify the response contains the correct line count
-    print('Step 2k: Verifying the response contains correct line count...')
-
-    # Get the actual line count from README.md
-    readme_lines = 157  # We know this from earlier: `wc -l README.md` = 157
-
-    # Look for the response in the chat messages
-    # The response should contain the number 157
-    message_selectors = [
-        f'text*="{readme_lines}"',
-        'text*="157"',
-        'text*="lines"',
-        '[data-testid="message"]',
-        '.message',
-        '[role="message"]',
-    ]
-
-    response_found = False
-    for selector in message_selectors:
-        try:
-            elements = page.locator(selector)
-            count = elements.count()
-            if count > 0:
-                print(f'Found {count} elements matching: {selector}')
-                # Check if any contain the line count
-                for i in range(count):
-                    element = elements.nth(i)
-                    if element.is_visible():
-                        text_content = element.text_content()
-                        if text_content and str(readme_lines) in text_content:
-                            print(
-                                f'Found response containing line count {readme_lines}: {text_content[:200]}...'
-                            )
-                            response_found = True
-                            break
-                if response_found:
-                    break
-        except Exception:
-            continue
-
-    if not response_found:
-        print(f'Response with line count {readme_lines} not found')
-        print('Checking all visible text on page...')
-
-        # Get all text content from the page - we need to use a selector
-        body_text = page.locator('body').text_content()
-        if str(readme_lines) in body_text:
-            print(f'Line count {readme_lines} found somewhere on the page')
-            response_found = True
-        else:
-            print(f'Line count {readme_lines} not found anywhere on the page')
-
-            # Look for any number that might be the line count
-            import re
-
-            numbers = re.findall(r'\b\d+\b', body_text)
-            print(f'Numbers found on page: {numbers[:20]}')  # Show first 20 numbers
-
-    # Final verification
-    if response_found:
-        print(
-            '✅ SUCCESS: Agent successfully completed the task and provided the correct line count!'
-        )
-    else:
-        print('❌ FAILURE: Agent response does not contain the expected line count')
-        # Don't fail the test completely, as the agent might have provided the answer in a different format
-        print('Test completed but response verification failed')
-
-    # Take a final screenshot
-    page.screenshot(path='test-results/12_final_result.png')
-    print('Screenshot saved: 12_final_result.png')
-
-    print('🎉 End-to-end test completed successfully!')
-
-
-if __name__ == '__main__':
-    """
-    Allow direct execution of this test file with command-line arguments.
-
-    Example usage:
-    - Run with visible browser:
-      python test_workflow.py --no-headless
-
-    - Run with visible browser and slow motion:
-      python test_workflow.py --no-headless --slow-mo=100
-
-    - Run a specific test:
-      python test_workflow.py --no-headless test_simple_browser_navigation
-    """
-    import sys
-
-    import pytest
-
-    # Default arguments
-    pytest_args = [
-        __file__,
-        '-v',
-        '--no-header',
-        '--capture=no',
-    ]
-
-    # Process command-line arguments
-    headless = True
-    slow_mo = 0
-    test_names = []
-
-    for arg in sys.argv[1:]:
-        if arg == '--no-headless':
-            headless = False
-        elif arg.startswith('--slow-mo='):
-            try:
-                slow_mo = int(arg.split('=')[1])
-            except (IndexError, ValueError):
-                print(f'Invalid slow-mo value: {arg}')
-                sys.exit(1)
-        elif arg.startswith('--'):
-            # Pass other pytest options directly
-            pytest_args.append(arg)
-        else:
-            # Assume it's a test name
-            test_name = arg
-            # Check if it's a function name without the file prefix
-            if not test_name.startswith('test_workflow.py::'):
-                test_name = f'{__file__}::{test_name}'
-            test_names.append(test_name)
-
-    # Add headless/no-headless flag
-    if headless:
-        pytest_args.append('--headless')
-    else:
-        pytest_args.append('--no-headless')
-
-    # Add slow-mo if specified
-    if slow_mo > 0:
-        pytest_args.append(f'--slow-mo={slow_mo}')
-
-    # Add test names if specified
-    if test_names:
-        pytest_args.extend(test_names)
-
-    # Print the command being run
-    print(f'Running: pytest {" ".join(arg for arg in pytest_args if arg != __file__)}')
-
-    # Run the tests
-    sys.exit(pytest.main(pytest_args))
